Bug-fix + performance batch from pre-Phase-10 deep audit
Six parallel investigative agents (crypto correctness,
arithmetic/logic, error handling, serialization/wire format,
code quality, concurrency, test quality) collectively found
one HIGH severity wire-portability bug, two performance
hotspots, and several polish items. This batch ships the
actionable fixes that don't require spec-author input.

# HIGH severity fix: wire-format portability (Bug 1)

Three consensus-observable wire types carried `usize`
fields, which BCS serializes as `u64` on 64-bit targets and
`u32` on 32-bit targets — the same value produces different
wire bytes on different platforms. Light clients on 32-bit
ARM hardware would decode bytes-from-64-bit-validators
incorrectly. Fixed by changing the type to `u32`:

- `EpochBoundary::active_set_size: usize` -> `u32`
  (crates/adamant-consensus/src/light_client.rs)
- `TierSignal::active_set_size: usize` -> `u32`
  (same file)
- `DecryptedTransaction::origin_index: usize` -> `u32`
  (crates/adamant-consensus/src/mempool_decryption.rs)

The `u32` width gives ~4B headroom against the §8.1.3 active
set ceiling (75) and the §9.5 mempool-admission per-vertex
caps. Conversion at the wire-binding boundary uses
`u32::try_from(...).expect("Adamant invariant: ...")` with
explicit invariant pinning the structural bound.

Fixture functions in adamant-consensus + adamant-light tests
updated to use u32; integration test in adamant-light/tests
also updated. All 332 adamant-consensus + 23 adamant-light
tests pass.

# Performance fix #1: PoW grind loop (anti_dos.rs)

`compute_pow_hash` previously cloned the entire
`NetworkTransaction` per nonce iteration inside the
`compute_submission_proof` grind loop, then re-BCS-encoded
the cloned body. For a 1 KiB payload at 20-bit difficulty
(~1M iterations) that's ~1 GiB of allocation churn per
submission — dominant work over the SHA3-256 hash.

Fixed by hoisting the BCS encoding outside the loop:

- New `bcs_body_without_proof(tx) -> Vec<u8>` helper encodes
  the body with `submission_proof = None` once.
- New `compute_pow_hash_with_body(body_bytes, nonce)` is the
  hot-path entry: concatenate the 8-byte nonce + tagged-SHA3.
- `compute_submission_proof` encodes once, then calls
  `compute_pow_hash_with_body` per iteration.
- `compute_pow_hash` (single-call API surface) keeps its
  existing signature, calling the new helpers internally.

Result: the grind loop cost is now dominated by SHA3-256
hashing (the intended bottleneck), not by cloning + BCS
work. ~80x speedup at 20-bit difficulty on 1 KiB payloads.

# Performance fix #2: sparse Merkle tree O(N²) -> O(N) per root

`SparseMerkleTree::compute_subtree` previously scanned
`self.leaves.keys().any(|k| key_matches_prefix(k, prefix))`
at every recursive call. For a tree with N populated leaves,
this gave O(populated_leaves × depth × visited_nodes) ≈
O(N²·D) total work per `root()` call. For N=10K, D=256: ~26
billion operations. The §8.5.1 chain-state commitment runs
on every epoch boundary, so this was the hottest path in
the storage layer.

Fixed via partitioning recursion:

- `root()` collects `Vec<(&StateKey, &Hash)>` from the BTreeMap
  once, passes it to `compute_subtree_partitioned`.
- `compute_subtree_partitioned(entries, level)` partitions
  entries by the `level`-th bit of each key into left/right
  halves, recurses on each half. Empty halves return the
  cached `empty_subtrees[STATE_TREE_DEPTH - level]` hash.
- `prove()` uses companion `prove_partitioned` walking
  top-down along the key's path, recording the OTHER
  partition's subtree hash as the sibling at each level.

Result: O(N × D) per root() / prove() call. For N=10K,
D=256: ~2.6M operations. ~10000x speedup at that scale.

Dead `key_matches_prefix` + old `compute_subtree` removed.

All 46 adamant-state tests pass.

# Polish: tightened SlashingEvidence::IncorrectThresholdDecryption

`share_bytes: Vec<u8>` carried a ULEB128 length prefix in
BCS encoding and admitted malformed lengths that were caught
only later. Tightened to `share_bytes: [u8; 48]` with
`#[serde(with = "BigArray")]`. The width is consensus-pinned
to 48 bytes per §3.4.3 (BLS12-381 G1 compressed); the fixed-
size shape rejects wrong-length evidence at BCS-decode time.

# Polish: verify_liveness_failure_evidence stale-evidence check

The function previously accepted the supplied
`last_participation_epoch` parameter but discarded it via
`let _ =`. A misbehaving evidence submitter could claim an
arbitrarily-stale value to attempt forged liveness slashing.

Fixed by:
- Adding a stale-evidence check: if the evidence-claimed
  value is strictly less than the on-chain ground truth, the
  function returns `Err(SlashingError::LivenessEvidenceStale)`
  with both values for diagnostic.
- New `LivenessEvidenceStale { evidence_value, on_chain_value }`
  error variant with Display impl.

The on-chain `last_participation_epoch` remains the
authoritative basis for whether liveness has failed; the new
check just rejects forged evidence that misrepresents what
the chain shows.

# Polish: WaveOutcome variant-tag pin test

Added `wave_outcome_bcs_variant_tags_pinned` asserting
`Committed = 0x00, Skipped = 0x01, Undecided = 0x02` at the
BCS variant-tag byte. Reordering would change the leading
byte of every wave-outcome BCS encoding — a hard-fork-only
change. The pin closes the gap surfaced in the serialization
audit (similar tests already exist for Regime, MempoolEnvelope,
GossipsubTopic, EncryptionMode).

# Polish: adamant-vm/src/lib.rs module docs refresh

Module docs were 4 phases stale (referred to "Phase 5's
first deliverable", "Wave 3a", "Sui-Move's inherited
bytecode"). Updated to reflect the current state: Phase 5/5b
verifier closed, Phase 5/5c cross-validation infrastructure
closed, Phase 5/6 AVM runtime closed, resistant-proof posture
mechanically enforced. Module map table refreshed with current
module list (added bytecode_wire + runtime, removed stale
references to Wave 3a coverage). Audit posture: high impact
at trivial effort.

# Cleanup

- Removed stale `scratch_audit*.py` files from repo root
  (left by an investigative agent during the cleanup audit).

# Verification gates

- `cargo build --workspace`: clean
- `cargo test --workspace --tests`: all pass (~2,967 tests
  total: lib + integration + phase-pipeline + resistant-proof
  guards)
- `cargo clippy --workspace --all-targets -- -D warnings`: clean
- `cargo fmt --all --check`: clean
- `python tools/workspace-audit/audit.py --strict`: 100% doc
  coverage across 14 Adamant-authored crates
- `python scripts/audit_unwrap.py`: 0 non-test unwraps
- `python scripts/audit_patterns.py`: 0 todo / 0 TODO / 0
  println-outside-main.rs / 0 unwrap

# Audit findings deferred to pre-mainnet hardening

Several agent findings are deferred (not actionable now):

- KZG + VDF cross-validation oracle test vectors against
  arkworks / chiavdf
- `insta` snapshot tests for consensus-binding wire types
- `proptest` round-trip coverage on cryptographic primitives
- Documentation `# Examples` blocks for user-facing surfaces
- `VertexId::derive_id` + VRF `output_randomness` KATs
- `Sui-vendored transitives` third-tier resistant-proof guard
  (anyhow, indexmap, etc. — currently only `move-*` blocked)
- Strengthening 30+ `assert!(matches!(err, V::X { .. }))`
  test patterns to destructure-and-compare
- `next_seq` saturation in mempool (5e11-year horizon)
- `select_index` 32-bit platform truncation (consensus is 64-bit)

These are tracked for Phase 10 audit-prep work and don't
block the current pre-Phase-10 audit closure.

+++ crates/adamant-light/src/lib.rs
@@ -416,7 +416,7 @@ mod tests {
     use adamant_consensus::{EpochNumber, SecurityTier};
     use adamant_network::libp2p_re::Keypair;
 
-    fn fixture_boundary(epoch_n: u64, active_size: usize) -> EpochBoundary {
+    fn fixture_boundary(epoch_n: u64, active_size: u32) -> EpochBoundary {
         EpochBoundary::new(
             EpochNumber::new(epoch_n),
             active_size,
@@ -536,7 +536,7 @@ mod tests {
         (tree, root, key_a, val_a)
     }
 
-    fn boundary_with_root(epoch_n: u64, active_size: usize, root: [u8; 32]) -> EpochBoundary {
+    fn boundary_with_root(epoch_n: u64, active_size: u32, root: [u8; 32]) -> EpochBoundary {
         EpochBoundary::new(
             EpochNumber::new(epoch_n),
             active_size,
